# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Operix API - The Operating System for Roblox Groups
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups/{groupId}/staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members for a group
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, flagged]
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaffMember
      tags: [staff]
      summary: Add a staff member
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffMemberBody"
      responses:
        "201":
          description: Created staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /groups/{groupId}/staff/{staffId}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get a staff member profile
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMemberDetail"
    patch:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffMemberBody"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaffMember
      tags: [staff]
      summary: Remove a staff member
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
        - name: staffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Sessions ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        robloxUsername:
          type: string
        robloxId:
          type: string
        rank:
          type: string
        role:
          type: string
          enum: [admin, hr, moderator, staff]
        status:
          type: string
          enum: [active, inactive, flagged, suspended]
        totalActivityMinutes:
          type: integer
        sessionsCount:
          type: integer
        warningsCount:
          type: integer
        casesCount:
          type: integer
        lastSeenAt:
          type: string
        joinedAt:
          type: string
        avatarUrl:
          type: string
      required: [id, groupId, robloxUsername, role, status, totalActivityMinutes, sessionsCount, warningsCount, casesCount, joinedAt]
    StaffMemberDetail:
      allOf:
        - $ref: "#/components/schemas/StaffMember"
        - type: object
          properties:
            recentSessions:
              type: array
              items:
                $ref: "#/components/schemas/Session"
            recentCases:
              type: array
              items:
                $ref: "#/components/schemas/Case"
            aiInsight:
              type: string
          required: [recentSessions, recentCases]
    CreateStaffMemberBody:
      type: object
      properties:
        robloxUsername:
          type: string
        robloxId:
          type: string
        rank:
          type: string
        role:
          type: string
          enum: [admin, hr, moderator, staff]
      required: [robloxUsername, role]
    UpdateStaffMemberBody:
      type: object
      properties:
        rank:
          type: string
        role:
          type: string
          enum: [admin, hr, moderator, staff]
        status:
          type: string
          enum: [active, inactive, flagged, suspended]

    # ── Sessions ─────────────────────────────────────────────────────────
    Session:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        staffId:
          type: integer
        staffUsername:
          type: string
        serverName:
          type: string
        startedAt:
          type: string
        endedAt:
          type: string
        durationMinutes:
          type: integer
        active:
          type: boolean
        actionsCount:
          type: integer
      required: [id, groupId, staffId, startedAt, active]
...
    Case:
      type: object
      properties:
        id:
          type: integer
        groupId:
          type: integer
        staffId:
          type: integer
        staffUsername:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, under_review, closed, appealed]
        autoGenerated:
          type: boolean
        aiSummary:
          type: string
        suggestedAction:
          type: string
        evidenceCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, groupId, title, severity, status, autoGenerated, createdAt]
...