# lib/api-zod/src/generated/types/staffMemberDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Operix API - The Operating System for Roblox Groups
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from "./case";
import type { Session } from "./session";
import type { StaffMember } from "./staffMember";

export type StaffMemberDetail = StaffMember & {
  recentSessions: Session[];
  recentCases: Case[];
  aiInsight?: string;
};